X64: Made hash computation in serializer accept 64-bit pointers. Also changed api Wrap function to only wrap suitably small pointers as Smis. Added Smi validity check and factory meethod for intptr_t.
Review URL: http://codereview.chromium.org/113023


git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@1875 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/serialize.h
@@ -70,7 +70,7 @@ class ExternalReferenceEncoder {
  private:
   HashMap encodings_;
   static uint32_t Hash(Address key) {
-    return reinterpret_cast<uint32_t>(key) >> 2;
+    return static_cast<uint32_t>(reinterpret_cast<uintptr_t>(key) >> 2);
   }
 
   int IndexOf(Address key) const;
